fsspec path handling
Proposed as fix for #3201. Some filesystems need the scheme as part of the path,
while others don't. FsspecStore.from_url() throws an exception if the scheme is
left in the path, for any filesystem except http and https. However, the swift fs
also needs the scheme in the path. This commit removes the exception, rather than
adding more special cases.